Move into its own file the external test
